🚚 Rename: ChatMessageList 컴포넌트 폴더 분리 및 적용 (#86)
